Update - Exclusão de arquivos desnecessários
