+++ KAUA_MATHEUS/primeiro_bimestre/calculadora.java
@@ -1,5 +1,12 @@
 package KAUA_MATHEUS.primeiro_bimestre;
 import java.util.Scanner;
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.File;
+import java.io.FileReader;
+import java.io.FileWriter;
+import java.io.IOException;
+import java.io.PrintWriter;
 import java.lang.Thread;
 
 public class calculadora {
@@ -55,7 +62,7 @@ public class calculadora {
 
                     // Criação de arquivo txt contendo o conteúdo
 
-                    Barquivo.criaArquivos("KAUA_MATHEUS/primeiro_bimestre/arquivoCompras.txt", "Qnt: " + 
+                    criaArquivos("KAUA_MATHEUS/primeiro_bimestre/arquivoCompras.txt", "Qnt: " + 
                     quantidade_produto + " -- Preço: R$" + 
                     preço_produto + " -- Desconto Total: " +
                     String.format("R$%.2f", (desc)) + "\n");
@@ -92,7 +99,7 @@ public class calculadora {
                     
                         int sair = 0;
                         
-                        Barquivo.lerTexto("KAUA_MATHEUS/primeiro_bimestre/arquivoCompras.txt");
+                        lerTexto("KAUA_MATHEUS/primeiro_bimestre/arquivoCompras.txt");
                         while (sair != 1) {
                             System.out.println("[1] - Sair");
                             sair = Integer.parseInt(scanner.nextLine());
@@ -146,6 +153,67 @@ public class calculadora {
         }
     
         public static void apagar(String caminho_Arquivo){
-            Barquivo.apagaArquivo(caminho_Arquivo);
+            apagaArquivo(caminho_Arquivo);
         }
+
+
+        public static void lerTexto(String pCaminhoArquivo){
+
+        int contador_linha = 0;
+
+        try(
+
+            BufferedReader buffReader = new BufferedReader(new FileReader(pCaminhoArquivo))
+        ){
+                String linha = "";
+
+                while(true){
+                    if(linha != null){
+                        System.out.println(linha);
+                        contador_linha += 1;
+                    }else{
+                        break;
+                    }
+                    linha = buffReader.readLine();
+                }
+                if(contador_linha > 0){
+                    System.out.println(contador_linha - 1);
+                }
+                else{
+                    System.out.println(0);
+                }
+                
+            }catch (IOException e){
+                e.printStackTrace();
+            }
+
+    }
+
+    public static void criaArquivos(String pCaminhoArquivo, String pTextoAEscrever){
+        try(
+
+            FileWriter criadorDeArquivos = new FileWriter(pCaminhoArquivo, true);
+            BufferedWriter buffer = new BufferedWriter(criadorDeArquivos);
+            PrintWriter escritorDeArquivos = new PrintWriter(buffer);
+        
+        ){
+            escritorDeArquivos.append(pTextoAEscrever);
+        }catch(IOException e){
+            e.printStackTrace();
+        }
+    }
+
+    public static void apagaArquivo(String pCaminhoArquivo){
+
+        File file = new File(pCaminhoArquivo);
+
+        if(file.delete()){
+            System.out.format("\u001B[31mO arquivo foi apagado com sucesso! \u001B[m");   
+        }
+
+        else{
+            System.out.println("\u001B[31mErro ao tentar apagar arquivo! \u001B[m");
+        }
+    }
+
     }
